cyber_record_reader/pyproject.toml
[tool.poetry]
name = "cyberrecordreader"
version = "0.1.0"
description = ""
authors = ["changnam hong <hcn1519@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
cyber-record = "^0.1.12"
pyyaml = "^6.0.1"
protobuf3-to-dict = "^0.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
